third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/loose-subtree-nth.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description("Element.querySelector matching nth-child of loose subtree.");

var root = document.createElement("div");
for (var i = 0; i < 12; i++) {
    var child = document.createElement("span");
    if (i == 11)
        child.id = "pass";
    root.appendChild(child);
}

shouldBe('root.querySelector(":nth-child(12)").id', '"pass"');
</script>
