third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/loose-subtree-nth-expected.txt

Element.querySelector matching nth-child of loose subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS root.querySelector(":nth-child(12)").id is "pass"
PASS successfullyParsed is true

TEST COMPLETE
